moved quality scripts folder
