+++ Assets/_Application/Quality/EnvironmentSettings.cs.meta
